+++ extensions/cornerstone/src/getPanelModule.tsx
@@ -51,19 +51,6 @@ const getPanelModule = ({ commandsManager, servicesManager, extensionManager }: 
     );
   };
 
-  const wrappedPanelMeasurement = ({ configuration }) => {
-    return (
-      <PanelMeasurement
-        commandsManager={commandsManager}
-        servicesManager={servicesManager}
-        extensionManager={extensionManager}
-        configuration={{
-          ...configuration,
-        }}
-      />
-    );
-  };
-
   return [
     {
       name: 'activeViewportWindowLevel',
@@ -76,7 +63,7 @@ const getPanelModule = ({ commandsManager, servicesManager, extensionManager }: 
       iconName: 'tab-linear',
       iconLabel: 'Measure',
       label: 'Measurement',
-      component: wrappedPanelMeasurement,
+      component: PanelMeasurement,
     },
     {
       name: 'panelSegmentation',