+++ platform/ui-next/src/components/index.ts
@@ -58,6 +58,7 @@ import { InputDialog, PresetDialog } from './OHIFDialogs';
 import { AboutModal, ImageModal, UserPreferencesModal } from './OHIFModals';
 import Modal from './Modal/Modal';
 import { FooterAction } from './FooterAction';
+import { InputFilter } from './InputFilter';
 
 import {
   DropdownMenu,
@@ -78,12 +79,18 @@ import {
 } from './DropdownMenu';
 import { Onboarding } from './Onboarding';
 import { DoubleSlider } from './DoubleSlider';
-import { DataRow } from './DataRow';
-import { MeasurementTable } from './MeasurementTable';
-import { SegmentationTable, useSegmentationTableContext } from './SegmentationTable';
+export { DataRow } from './DataRow';
+export { MeasurementTable } from './MeasurementTable';
+import {
+  SegmentationTable,
+  useSegmentationTableContext,
+  useSegmentationExpanded,
+  useSegmentStatistics,
+} from './SegmentationTable';
 import { Toaster, toast } from './Sonner';
 import { StudySummary } from './StudySummary';
 import { ErrorBoundary } from './Errorboundary';
+export * from './ColorCircle';
 import { Header } from './Header';
 import { Card, CardHeader, CardFooter, CardTitle, CardDescription, CardContent } from './Card';
 import {
@@ -106,6 +113,9 @@ import {
   ToolButtonListDivider,
 } from './ToolButton';
 import { ToolSettings } from './OHIFToolSettings';
+
+// Segmentation Context Exports
+export { useSegmentationTableContext, useSegmentationExpanded, useSegmentStatistics };
 
 export {
   Numeric,
@@ -199,12 +209,9 @@ export {
   TooltipTrigger,
   TooltipContent,
   TooltipProvider,
-  DataRow,
-  MeasurementTable,
   Toaster,
   toast,
   SegmentationTable,
-  useSegmentationTableContext,
   StudySummary,
   Header,
   Card,
@@ -237,4 +244,5 @@ export {
   UserPreferencesModal,
   FooterAction,
   ToolSettings,
+  InputFilter,
 };