+++ platform/ui-next/src/utils/index.ts
@@ -1,3 +1,4 @@
 import { getToggledClassName } from './getToggledClassName';
+import roundNumber from './roundNumber';
 
-export { getToggledClassName };
+export { getToggledClassName, roundNumber };